client/requirements.txt
psutil
requests
pyautogui
pyperclip
Pillow
pynput
swinlnk
websockets
httpx
pydantic
scapy